数据表生成代码修改，common工具包修改，添加产品方法若干

+++ iccp-milktea/iccp-milktea-mapper/src/main/java/com/example/milktea/mapper/ProductStapleDOMapper.java
@@ -11,7 +11,7 @@ public interface ProductStapleDOMapper {
 
     int deleteByExample(ProductStapleDOExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(ProductStapleDO record);
 
@@ -19,7 +19,7 @@ public interface ProductStapleDOMapper {
 
     List<ProductStapleDO> selectByExample(ProductStapleDOExample example);
 
-    ProductStapleDO selectByPrimaryKey(Integer id);
+    ProductStapleDO selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") ProductStapleDO record, @Param("example") ProductStapleDOExample example);
